init-crud

+++ src/Controller/TaskController.php
@@ -4,12 +4,10 @@ namespace App\Controller;
 
 use App\Entity\Task;
 use App\Form\TaskType;
-use App\Repository\TaskRepository;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 /**
  * @Route("/task")
@@ -18,10 +16,12 @@ class TaskController extends AbstractController
 {
     /**
      * @Route("/", name="task_index", methods={"GET"})
+     * @param UserInterface $userLogged
+     * @param Request $request
+     * @return Response
      */
-    public function index(TaskRepository $taskRepository, UserInterface $userLogged, Request $request): Response
+    public function index(UserInterface $userLogged, Request $request): Response
     {
-        $page = $request->query->getInt('page', 1);
         $tasks = $userLogged->getTasks();
 
         return $this->render('task/index.html.twig', [
@@ -31,6 +31,9 @@ class TaskController extends AbstractController
 
     /**
      * @Route("/new", name="task_new", methods={"GET","POST"})
+     * @param Request $request
+     * @param UserInterface $userLogged
+     * @return Response
      */
     public function new(Request $request, UserInterface $userLogged): Response
     {
@@ -61,9 +64,14 @@ class TaskController extends AbstractController
 
     /**
      * @Route("/{id}", name="task_show", methods={"GET"})
+     * @param Task $task
+     * @param UserInterface $userLogged
+     * @return Response
      */
-    public function show(Task $task): Response
+    public function show(Task $task, UserInterface $userLogged): Response
     {
+        if($userLogged->getId() !== $task->getUser()->getId()) return $this->redirectToRoute('task_index');
+
         return $this->render('task/show.html.twig', [
             'task' => $task,
         ]);
@@ -71,9 +79,15 @@ class TaskController extends AbstractController
 
     /**
      * @Route("/{id}/edit", name="task_edit", methods={"GET","POST"})
+     * @param Request $request
+     * @param Task $task
+     * @param UserInterface $userLogged
+     * @return Response
      */
-    public function edit(Request $request, Task $task): Response
+    public function edit(Request $request, Task $task, UserInterface $userLogged): Response
     {
+        if($userLogged->getId() !== $task->getUser()->getId()) return $this->redirectToRoute('task_index');
+
         $form = $this->createForm(TaskType::class, $task);
         $form->handleRequest($request);
 
@@ -91,9 +105,15 @@ class TaskController extends AbstractController
 
     /**
      * @Route("/{id}", name="task_delete", methods={"DELETE"})
+     * @param Request $request
+     * @param Task $task
+     * @param UserInterface $userLogged
+     * @return Response
      */
-    public function delete(Request $request, Task $task): Response
+    public function delete(Request $request, Task $task, UserInterface $userLogged): Response
     {
+        if($userLogged->getId() !== $task->getUser()->getId()) return $this->redirectToRoute('task_index');
+
         if ($this->isCsrfTokenValid('delete'.$task->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($task);
